feat: organize constants and types/update docs
